+++ lib/active_storage_validations/shared/asv_errorable.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative "../asv_attachable_adapter"
+
 module ActiveStorageValidations
   module ASVErrorable
     extend ActiveSupport::Concern
@@ -34,15 +36,7 @@ module ActiveStorageValidations
     end
 
     def filename_from(file)
-      case file
-      when ActiveStorage::Attached, ActiveStorage::Attachment then file.blob&.filename
-      when ActiveStorage::Blob then file.filename
-      when ActionDispatch::Http::UploadedFile, Rack::Test::UploadedFile then file.original_filename
-      when String then ActiveStorage::Blob.find_signed!(file)&.filename
-      when Hash then file[:filename]
-      when File then File.basename(file)
-      when Pathname then File.basename(file)
-      end
+      ASVAttachableAdapter.filename_for(file)
     end
   end
 end